Create README.md
Exploratory Data Analysis on Dataset - Terrorism
This project aims to perform exploratory data analysis on a dataset of global terrorist attacks from 1970 to 2017. The dataset contains information about the location, date, target, weapon, group, and casualties of each attack. The project also seeks to answer the question: What are the hot zones of terrorism?

Key Findings
The hottest zone of terrorism is Middle East and North Africa, with a total of 158764 attacks in the dataset.
The year 2014 witnessed the highest number of terrorist attacks, with 57236 incidents across the globe.
The most affected targets by these terror attacks are private citizens and property, with a count of 609154 victims.
Afghanistan is the most attacked country, with 45334 attacks, followed by Iraq, Pakistan, and India.
The most common weapon type used by the terrorists is bombs and explosives, with 264765 attacks, followed by firearms, melee, and incendiary.
The most notorious group or trouble making group is of Abu Sayyaf Group (ASG), which is responsible for 1275 attacks, mostly in the Philippines.
Tools Used
Microsoft Power BI: A business analytics tool that allows users to create interactive dashboards and reports.
Power Query: A data transformation and mashup tool that enables users to connect, combine, and shape data from various sources.
How to Run the Project
Download the dataset from this link and save it as globalterrorismdb_0718dist.csv in your local directory.
Download the Power BI Desktop application from this link and install it on your computer.
Open the Power BI Desktop application and click on Get Data. Select Text/CSV and browse to the location of the dataset file. Click on Load.
Click on Edit Queries to open the Power Query Editor. Here you can apply various transformations and filters to the data as per your needs. For example, you can remove unnecessary columns, rename columns, change data types, etc.
Click on Close & Apply to return to the Power BI report view. Here you can create various visualizations and charts to explore the data and discover insights. For example, you can use maps, bar charts, line charts, pie charts, etc.
To save your work, click on File and then Save. Give a name to your Power BI file and choose a location to save it. The file will have a .pbix extension.
To share your work, you can either publish your report to the Power BI service or export it as a PDF file. For more information, see this link.